requirements.txt
pandas~=2.3.3
numpy~=1.26.4
omegaconf~=2.3.0
pillow~=12.0.0
torchvision~=0.15.2
matplotlib~=3.10.6
matplotlib-inline~=0.2.1
scikit-learn~=1.7.2